libretro-common/vfs: cast write() result for -1 / size checks
The two -Wsign-compare warnings in retro_vfs_file_write_impl
both stem from comparing a size_t-typed 'result' against -1
(write() failure sentinel that wraps to SIZE_MAX in unsigned
arithmetic), and against an int64_t stream->size after an
addition that promotes pos to unsigned.

Cast 'result' to ssize_t for the failure sentinel check and to
int64_t for the size addition.  Both casts are safe: write()
never returns a value larger than ssize_t can hold positively,
and stream->size / pos are int64_t.

This is a local patch to vendored libretro-common; should be
upstreamed to the libretro-common repo on next sync.

+++ libretro/libretro-common/vfs/vfs_implementation.c
@@ -726,7 +726,7 @@ int64_t retro_vfs_file_write_impl(libretro_vfs_implementation_file *stream, cons
       pos    = retro_vfs_file_tell_impl(stream);
       result = fwrite(s, 1, (size_t)len, stream->fp);
 
-      if (result != -1 && pos + result > stream->size)
+      if ((ssize_t)result != -1 && pos + (int64_t)result > stream->size)
          stream->size = pos + result;
 
       return result;
@@ -739,7 +739,7 @@ int64_t retro_vfs_file_write_impl(libretro_vfs_implementation_file *stream, cons
    pos    = retro_vfs_file_tell_impl(stream);
    result = write(stream->fd, s, (size_t)len);
 
-   if (result != -1 && pos + result > stream->size)
+   if ((ssize_t)result != -1 && pos + (int64_t)result > stream->size)
       stream->size = pos + result;
 
    return result;
